Flush before resetting colors
On Windows the write of the text and the end characters may be buffered
together before being printed so that setting colors appears to have no
effect since colors are then reset before the last flush. This ensures
that colors are set, the text flushed, then colors are reset and then
the end character (which should not be affected by colors) is flushed.